Code review changes. Renamed utils module to _utils
