+++ src/fixpoint/_utils/__init__.py
@@ -1,0 +1,5 @@
+"""
+This module contains the utility functions and classes.
+"""
+
+from .completions import *
