+++ src/fixpoint/_utils/completions.py
