--- src/fixpoint/utils/__init__.py
@@ -1,7 +1,0 @@
-"""
-This module contains the utility functions and classes.
-"""
-
-from .completions import decorate_instructor_completion_with_fixp
-
-__all__ = ["decorate_instructor_completion_with_fixp"]
